refactor: move packages to the correct folder
