--- Api/csharp/ArmoniK.Api.Common/Options/GrpcChannel.cs
@@ -1,40 +1,0 @@
-// This file is part of the ArmoniK project
-// 
-// Copyright (C) ANEO, 2021-2022. All rights reserved.
-//   W. Kirschenmann   <wkirschenmann@aneo.fr>
-//   J. Gurhem         <jgurhem@aneo.fr>
-//   D. Dubuc          <ddubuc@aneo.fr>
-//   L. Ziane Khodja   <lzianekhodja@aneo.fr>
-//   F. Lemaitre       <flemaitre@aneo.fr>
-//   S. Djebbar        <sdjebbar@aneo.fr>
-//   J. Fonseca        <jfonseca@aneo.fr>
-// 
-// Licensed under the Apache License, Version 2.0 (the "License")
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-// 
-//         http://www.apache.org/licenses/LICENSE-2.0
-// 
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-using JetBrains.Annotations;
-
-namespace ArmoniK.Api.Common.Options;
-
-[PublicAPI]
-public class GrpcChannel
-{
-  /// <summary>
-  ///   Address or path of the resource used to communicate for this Grpc Channel
-  /// </summary>
-  public string Address { get; set; } = "/tmp/armonik.sock";
-
-  /// <summary>
-  ///   Type of Grpc Socket used
-  /// </summary>
-  public GrpcSocketType SocketType { get; set; } = GrpcSocketType.UnixDomainSocket;
-}
