--- Api/csharp/ArmoniK.Api.Worker/ArmoniK.Api.Worker.csproj
@@ -1,51 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk.Web">
-
-  <PropertyGroup>
-    <TargetFrameworks>net6.0</TargetFrameworks>
-    <Company>ANEO</Company>
-    <Copyright>Copyright (C) ANEO, 2021-2022</Copyright>
-    <PackageLicenseExpression>Apache-2.0</PackageLicenseExpression>
-    <PackageRequireLicenseAcceptance>True</PackageRequireLicenseAcceptance>
-    <PackageOutputPath>../publish</PackageOutputPath>
-    <LangVersion>10</LangVersion>
-    <Nullable>enable</Nullable>
-    <TargetType>library</TargetType>
-    <IsPackable>true</IsPackable>
-    <Optimize>true</Optimize>
-    <DebugType>full</DebugType>
-    <IncludeSymbols>true</IncludeSymbols>
-    <SymbolPackageFormat>snupkg</SymbolPackageFormat>
-    <AssemblyOriginatorKeyFile>../kp.snk</AssemblyOriginatorKeyFile>
-  </PropertyGroup>
-
-  <PropertyGroup Condition=" '$(Configuration)' == 'Debug' ">
-    <DefineConstants>DEBUG;TRACE</DefineConstants>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Grpc.Net.Client" Version="2.49.0" />
-    <PackageReference Include="JetBrains.Annotations" Version="2022.1.0" />
-    <PackageReference Include="System.Linq.Async" Version="6.0.1" />
-    <PackageReference Include="Grpc.AspNetCore.Server.Reflection" Version="2.49.0" />
-    <PackageReference Include="Serilog.Extensions.Hosting" Version="5.0.1" />
-    <PackageReference Include="Serilog.Formatting.Compact" Version="1.1.0" />
-    <PackageReference Include="Serilog.AspNetCore" Version="6.0.1" />
-    <PackageReference Include="Grpc.Tools" Version="2.49.1">
-      <PrivateAssets>all</PrivateAssets>
-      <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
-    </PackageReference>
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\ArmoniK.Api.Common.Channel\ArmoniK.Api.Common.Channel.csproj" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <Protobuf Include="..\..\..\Protos\V1\worker_service.proto" GrpcServices="Server">
-      <Link>gRPC\Protos\worker_service.proto</Link>
-    </Protobuf>
-    <Protobuf Include="..\..\..\Protos\V1\agent_service.proto" GrpcServices="Client">
-      <Link>gRPC\Protos\agent_service.proto</Link>
-    </Protobuf>
-  </ItemGroup>
-</Project>
