+++ packages/csharp/ArmoniK.Api.Client/ArmoniK.Api.Client.csproj
