+++ packages/csharp/ArmoniK.Api.Common/Utils/Disposable.cs
@@ -1,0 +1,48 @@
+// This file is part of the ArmoniK project
+// 
+// Copyright (C) ANEO, 2021-2022. All rights reserved.
+//   W. Kirschenmann   <wkirschenmann@aneo.fr>
+//   J. Gurhem         <jgurhem@aneo.fr>
+//   D. Dubuc          <ddubuc@aneo.fr>
+//   L. Ziane Khodja   <lzianekhodja@aneo.fr>
+//   F. Lemaitre       <flemaitre@aneo.fr>
+//   S. Djebbar        <sdjebbar@aneo.fr>
+//   J. Fonseca        <jfonseca@aneo.fr>
+// 
+// Licensed under the Apache License, Version 2.0 (the "License")
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+// 
+//         http://www.apache.org/licenses/LICENSE-2.0
+// 
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+
+namespace ArmoniK.Api.Common.Utils;
+
+internal static class Disposable
+{
+  public static IDisposable Empty = Create(() =>
+                                           {
+                                           });
+
+  public static IDisposable Create(Action action)
+    => new DisposableImpl(action);
+
+  private sealed class DisposableImpl : IDisposable
+  {
+    private readonly Action action_;
+
+    public DisposableImpl(Action action)
+      => action_ = action;
+
+    /// <inheritdoc />
+    public void Dispose()
+      => action_();
+  }
+}
