+++ packages/csharp/ArmoniK.Api.Core/ArmoniK.Api.Core.csproj
